+++ classes/Cache/Version/v52/class.xlvoCache.php
@@ -100,7 +100,7 @@ class xlvoCache extends \ilGlobalCache implements xlvoCacheService, Initialisabl
     {
         try
         {
-            return xlvoConf::getConfig(xlvoConf::F_USE_GLOBAL_CACHE) === 1;
+            return (int)xlvoConf::getConfig(xlvoConf::F_USE_GLOBAL_CACHE) === 1;
         }
         catch (ErrorException $exceptionex) //catch exception while dbupdate is running. (xlvoConf is not ready at that time).
         {
